New branch
Added config.json instead of command line arguments
program now starts from main.py instead of collector.py
sources are in folder src
