--- ClassDiagram.puml
@@ -1,73 +1,0 @@
-@startuml
-'https://plantuml.com/class-diagram
-title ITR project Class Diagram
-class Collector{
--SCALE : int
-#scheduler : Scheduler
-#draw : Drawer
-+ppcm()
-+get_wcet()
-+arg_parser()
-}
-
-class Task{
-# name : String
-#period : int
-#deadline : int
-#wcet : int
-#executed_time : int
-#time : int
-
-+set_wcet()
-+display()
-}
-
-class Drawer{
--ECHELLE : int
--color : List[str]
-#cases : List[Case]
-#perdiods : List[int]
-#hyperperiod : int
-
-+draw_schedule()
-}
-
-class Scheduler{
-    #tasks :List[Tasks]
-    #hyperperiod : int
-    #time : int
-    #EDF : List[Case]
-    #RM : List[Case]
-
-    +schedule_preempt()
-    +schedule_non_preempt()
-    +RM_feasibility()
-    +EDF_feasibility()
-    *assign_RM_prio()
-    *assign_EDF_prio()
-    *get_next_RM_function()
-    *get_next_EDF_function()
-    *get_next_deadline()
-    *get_level()
-}
-object Case {
-    #name : str
-    #begin_time : int
-    #end_time : int
-    #level : int
-}
-
-object HEPTANE{
-    xml attributes
-}
-object SVG {
-    svg attributes
-}
-Collector <|-- Scheduler
-Collector <|-- Drawer
-Scheduler <|--* Task
-Scheduler <|--* Case
-
-HEPTANE --> Collector
-Collector -> SVG
-@enduml
